Move interfaces to unmock-core.
